Move place.js
