+++ server/lib/place.js
@@ -6,7 +6,7 @@ var Promise = require("bluebird");
 var config = require("../../config/config.js");
 
 /**
- * This model encapsulates Google Places functionality.
+ * This service encapsulates Google Places functionality.
  * 
  * @return {function}
  */
@@ -58,20 +58,20 @@ module.exports = function() {
     /**
      * Generate a request for the distance between two Google Places IDs.
      * 
-     * @param  {string} placeId1    A valid Google Places ID
-     * @param  {string} placeId2    A valid Google Places ID
+     * @param  {string} originPlaceId    A valid Google Places ID
+     * @param  {string} destPlaceId    A valid Google Places ID
      * 
      * @return {object}  The distance request object
      */
-    model.distance = (placeId1, placeId2) => {
+    model.distance = (originPlaceId, destPlaceId) => {
         var maps = new GoogleMaps.createClient({
             key: config.google.apiKey,
             Promise: Promise
         });
         
         return maps.distanceMatrix({
-            origins: "place_id:" + placeId1,
-            destinations: "place_id:" + placeId2
+            origins: "place_id:" + originPlaceId,
+            destinations: "place_id:" + destPlaceId
         });
     };
